requirements.txt
aiogram==3.3.0
langchain==0.1.5
chromadb==0.4.22
google-generativeai==0.3.2
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1
numpy==1.24.3
pandas==2.0.3
python-multipart==0.0.6